pyproject.toml
[tool.pdm]
distribution = true
[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
[project]
name = "livraria_django"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "asgiref==3.8.1",
    "attrs==24.2.0",
    "blinker==1.8.2",
    "brotli==1.1.0",
    "certifi==2024.8.30",
    "cffi==1.17.1; platform_python_implementation != \"PyPy\"",
    "cfgv==3.4.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloudinary==1.41.0",
    "colorama==0.4.6; sys_platform == \"win32\" or platform_system == \"Windows\"",
    "cryptography==43.0.1",
    "distlib==0.3.8",
    "dj-database-url==2.2.0",
    "django>=5.1.3",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==4.4.0",
    "django-extensions==3.2.3",
    "django-filter==24.3",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.27.2",
    "filelock==3.16.0",
    "flask==3.0.3",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "identify==2.6.0",
    "idna==3.8",
    "inflection==0.5.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "markupsafe==2.1.5",
    "netifaces==0.11.0",
    "nodeenv==1.9.1",
    "packaging==24.1",
    "passage-identity==2.1.4",
    "pillow==10.4.0",
    "platformdirs==4.3.2",
    "pre-commit==3.8.0",
    "psycopg2-binary==2.9.9",
    "pycparser==2.22; platform_python_implementation != \"PyPy\"",
    "pydotplus==2.0.2",
    "pyjwt==2.9.0",
    "pyparsing==3.1.4",
    "python-dotenv==1.0.1",
    "python-magic==0.4.27; sys_platform == \"linux\"",
    "python-magic-bin==0.4.14; sys_platform == \"win32\"",
    "pyyaml==6.0.2",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rpds-py==0.20.0",
    "setuptools==74.1.2",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1; sys_platform == \"win32\"",
    "uritemplate==4.1.1",
    "urllib3==2.2.2",
    "uvicorn==0.30.6",
    "virtualenv==20.26.4",
    "werkzeug==3.0.4",
    "whitenoise[brotli]==6.7.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
